requirements.txt
amqp==5.0.9
anyjson==0.3.3
arabic-reshaper==2.1.3
asgiref==3.4.1
billiard==3.6.4.0
Brotli==1.0.9
cached-property==1.5.2
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cssselect2==0.4.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==2.0.0
Django==3.2.11
django-betterforms==1.2
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-decouple==2.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
fonttools==4.27.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.3
itypes==1.2.0
Jinja2==3.0.3
kombu==5.1.0
lxml==4.8.0
MarkupSafe==2.0.1
numpy==1.19.5
openapi-codec==1.3.2
packaging==21.3
pandas==1.1.5
pikepdf==3.2.0
Pillow==8.4.0
prompt-toolkit==3.0.28
psycopg2==2.7.4
pycparser==2.21
pydyf==0.1.2
PyJWT==2.3.0
pyparsing==3.0.7
PyPDF2==1.26.0
pyphen==0.11.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2021.3
redis==4.1.4
reportlab==3.6.5
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.13.3
xhtml2pdf==0.2.5
xlrd==2.0.1
zipp==3.6.0
zopfli==0.1.9
